test: extract assertFormLayoutGrid test helper
