+++ packages/form-layout/test/form-layout-auto-responsive.test.js
@@ -3,30 +3,7 @@ import { fixtureSync, nextFrame, nextResize } from '@vaadin/testing-helpers';
 import '../src/vaadin-form-layout.js';
 import '../src/vaadin-form-item.js';
 import '../src/vaadin-form-row.js';
-
-function assertFormLayoutGrid(layout, { columns, rows }) {
-  const children = [...layout.children];
-  expect(new Set(children.map((child) => child.offsetLeft)).size).to.equal(columns, `expected ${columns} columns`);
-  expect(new Set(children.map((child) => child.offsetTop)).size).to.equal(rows, `expected ${rows} rows`);
-}
-
-function assertFormLayoutLabelPosition(layout, { position }) {
-  const columnWidth = parseFloat(layout.columnWidth);
-  const labelWidth = parseFloat(getComputedStyle(layout).getPropertyValue('--vaadin-form-layout-label-width'));
-  const labelSpacing = parseFloat(getComputedStyle(layout).getPropertyValue('--vaadin-form-layout-label-spacing'));
-
-  [...layout.children].forEach((child) => {
-    if (child.localName !== 'vaadin-form-item') {
-      return;
-    }
-
-    if (position === 'aside') {
-      expect(child.offsetWidth).to.equal(columnWidth + labelWidth + labelSpacing, 'expected labels to be aside');
-    } else {
-      expect(child.offsetWidth).to.equal(columnWidth, 'expected labels to be above');
-    }
-  });
-}
+import { assertFormLayoutGrid, assertFormLayoutLabelPosition } from './helpers.js';
 
 describe('form-layout auto responsive', () => {
   let container, layout, fields;
